% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Badanie gry przy zadanej strategii}
\usage{
badaj_gre(strategia, ile = 1000)
}
\arguments{
\item{strategia}{Strategia, przy ktorej ma byc badana gra}
}
\value{
Funkcja zwraca nastepujace statystyki:
minimum, 1. kwartyl, mediana, srednia arytmetyczna, 3. kwartyl, maksimum
}
\description{
Funkcja wywolujaca funkcje \code{gra()} 10000 razy przy zadanej strategii i zwracajaca podstawowe statystyki
}
\examples{
wyniki <- badaj_gre(strategia)

}
\author{
Agnieszka Ciepielewska
}

