% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_change_lists.R
\name{generate_change_lists}
\alias{generate_change_lists}
\title{Generowanie listy mozliwych wymian zwierzecia na mniejsze i mniejszych na jedno duze}
\usage{
generate_change_lists(prices_sorted)
}
\arguments{
\item{prices_sorted}{Ceny zwierzat, posortowane malejaca}
}
\value{
lista list macierzy wymian
}
\description{
Funkcja \code{generate_change_lists()} generuje dwie listy: wszystkie mozliwe wymiany zwierzecia na mniejsze 
i wszytskie mozliwe wymiany malych zwierza na jedno duze, wykorzystuje
funkcje \code{generate_exchanges()}
}
\examples{
wymiany_na_wieksze <- generate_change_lists[[1]]
wymiany_na_mniejsze <- generate_change_lists[[2]]

}
\author{
Agnieszka Ciepielewska
}

