% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_exchanges.R
\name{generate_exchanges}
\alias{generate_exchanges}
\title{Generowanie mozliwych wymian zwierzecia na mniejsze}
\usage{
generate_exchanges(n_rabbits, prices_sorted, pre_moves = c(), moves = c())
}
\arguments{
\item{n_rabbits}{Cena zwierzecia, ktore chcemy rozbic (w krolikach)}

\item{prices_sorted}{Ceny zwierzat, posortowane malejaca}

\item{pre_moves}{Wymiana, ktora juz nastapila}

\item{moves}{To, co chcemy jeszcze wyznaczyc}
}
\value{
moves Macierz wszystkich mozliwych wymian zadanego zwierzecia na tansze
}
\description{
Funkcja \code{generate_exchanges()} generuje wszystkie mozliwe wymiany zwierzecia na mniejsze, wykorzystywana w 
funkcji \code{generate_change_lists()}
}
\examples{
n_rabbits <- prices_sorted("kon")
mozliwe_wymiany <- generate_exchanges(n_rabbits, prices_sorted)

}
\author{
Agnieszka Ciepielewska
}

