% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Jedna gra w Super Farmera}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{Strategia, ktora ma byc wykorzystywana do grania}
}
\value{
Liczba tur, ktora byla potrzebna na wygranie gry przy tej strategii
}
\description{
Funkcja \code{gra()} rozgrywa jedna gre w Super Farmera przy zadanej strategii i zwraca liczbe tur, 
ktora byla potrzebna, aby gracz wygral
}
\examples{
\dontrun{
wynik <- gra(strategia)
}

}
\author{
Agnieszka Ciepielewska
}

