% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moves_generation.R
\name{moves_generation}
\alias{moves_generation}
\title{Generowanie mozliwych rochow}
\usage{
moves_generation(stock_status)
}
\arguments{
\item{stock_status}{Aktualny stan stada gracza}
}
\value{
possible_moves Macierz wszystkich mozliwych ruchow
}
\description{
Funkcja \code{moves_generation()} wszytkie mozliwe ruchy gracza przy danym stadzie.
Wykorzystuje funkcje \code{generate_change_lists()}
}
\examples{
stan_stada <- c(2,4,1,0,0,0,0)
stan_stada <- moves_generation(stan_stada)

}
\author{
Agnieszka Ciepielewska
}

