% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{reproduction}
\alias{reproduction}
\title{Rozmnazanie zwierzat po rzucie}
\usage{
reproduction(die1_result, die2_result, stock_status, max_stock)
}
\arguments{
\item{die1_result}{String, bedacy wynikiem rzutu pierwsza kostka (tej z wilkiem)}

\item{die2_result}{String, bedacy wynikiem rzutu druga kostka (tej z lisem)}

\item{stock_status}{Wektor posiadanych przez gracza zwierzat}

\item{max_stock}{Wektor maksymalnej liczby zwierzat w stadzie}
}
\value{
stock_status Stan stada gracza po rozpatrzeniu rzutow kostkami
}
\description{
Funkcja przeprowadzajaca zmiany w stadzie gracza po rzucie kostkami, 
wykorzystuje funkcje \code{update_stocku}
}
\examples{
stock_status <- reproduction(die1_result="krowa", die2_result="krolik", stock_status=c(0,1,0,2,0,0,0), 
max_stock=c(60,24,20,12,6,2,4))

}
\author{
Agnieszka Ciepielewska
}

