% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_EV.R
\name{strategia_EV}
\alias{strategia_EV}
\title{Strategia Expected Value}
\usage{
strategia_EV(stock_status, max_stock = c(krolik = 60, owca = 24, swinia = 20,
  krowa = 12, kon = 6, maly_pies = 4, duzy_pies = 2))
}
\arguments{
\item{stock_status}{Stan stada gracza przed ruchem}

\item{max_stock}{Stan stada glownego na poczatku gry.}

\item{max_stock}{Stan stada glownego na poczatku gry.}
}
\value{
Stan stada gracza po ruchu
}
\description{
Strategia polega na wybraniu ruchu, dla ktorego wartosc
 oczekiwana liczona w krolikach jest najwieksza. Funkcja \code{moves_generation()} 
 generuje wszystkie mozliwe ruchy. Funkcja \code{get_the_best_state} zwraca ruch o 
 największej
}
\author{
Marek Wawreniuk
}

