% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_HEV.R
\name{strategia_HEV}
\alias{strategia_HEV}
\title{Strategia Horse Expected Value}
\usage{
strategia_HEV(stock_status, max_stock = c(krolik = 60, owca = 24, swinia = 20,
  krowa = 12, kon = 6, maly_pies = 4, duzy_pies = 2))
}
\arguments{
\item{stock_status}{Stan stada gracza przed ruchem}

\item{max_stock}{Stan stada glownego na poczatku gry.}
}
\value{
Stan stada gracza po ruchu
}
\description{
Taktyka bedaca malym rozszerzeniem \code{strategia_EV()}. Uwzglednia ona to, ze poprzednia
 nie bierze pod uwage tego, ze wygrana partii nastepuje po tym jak zdobedziemy wszystkie zwierzeta, nie
po tym jak bedziemy mieli najwieksza wartosc stada w krolikach. Dlatego po pierwsze skupiamy sie na 
zakupie reszty zwierzat jesli mamy dwa konie. Potem na zakupie konia jesli nast tylko na to stac i
konia nie mamy, nastepnie jesli stac nas na wygranie gry, to kupujemy wszystkie zwierzeta po kolei, 
 ale w taki sposob zeby nie pozbywac sie innych (zeby nie wyzerowac jednego kosztem innego). Jesli 
zaden z powyzszych warunkow nie zachodzi, to wracamy do strategii \code{strategia_EV}.
}
\author{
Marek Wawreniuk
}

