% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttheme_kukurydza.R
\name{ttheme_kukurydza}
\alias{ttheme_kukurydza}
\title{Wyglad tabeli}
\usage{
ttheme_kukurydza()
}
\value{
tt1 szablon wygladu
}
\description{
Funkcja \code{ttheme_kukurydza} generuje ttheme do tabeli statystyk
}

