% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stocku.R
\name{update_stocku}
\alias{update_stocku}
\title{Aktualizacja stada}
\usage{
update_stocku(stock_status, max_stock, die1_result = "", die2_result = "")
}
\arguments{
\item{stock_status}{Wektor posiadanych przez gracza zwierzat}

\item{max_stock}{Wektor maksymalnej liczby zwierzat w stadzie}

\item{die1_result}{String, bedacy wynikiem rzutu pierwsza kostka (tej z wilkiem), domyslnie pusty string}

\item{die2_result}{String, bedacy wynikiem rzutu druga kostka (tej z lisem), domyslnie pusty string}
}
\value{
Funkcja zwraca zaktualizowany stan stada gracza
}
\description{
Funkcja pomocnicza wykorzystywana przez funkcje \code{reproduction()} do aktualizowania stanu stada gracza
}
\examples{
\dontrun{
stock_status <- update_stocku(stock_status=c(0,1,0,2,0,0,0), 
max_stock=c(60,24,20,12,6,2,4), die1_result="krowa")
}

}
\author{
Agnieszka Ciepielewska
}

