% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka_KukurydzaNaPizzy.R
\name{wizytowka_KukurydzaNaPizzy}
\alias{wizytowka_KukurydzaNaPizzy}
\title{Wizytowka}
\usage{
wizytowka_KukurydzaNaPizzy(strategia)
}
\arguments{
\item{strategia}{Strategia, dla ktorej na byc wygenerowana wizytowka}
}
\value{
Wizytowka.pdf Wizytowka, pokazujaca statystyki na temat strategii oraz jej dzialanie
}
\description{
Funkcja \code{wizytowka_KukurydzaNaPizzy} generuje wizytowke wybranej 
strategii w formacie pdf
}

