% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Badanie gry przy zadanej strategii}
\usage{
badaj_gre(strategia, ile = 10000)
}
\arguments{
\item{strategia}{Strategia, przy ktorej ma byc badana gra}

\item{ile}{Ile razy fukcja ma zagrac w SuperFarmera przy zadanej strategii, domyslnie 10000}
}
\value{
Funkcja zwraca wektor liczby ruchów w kolejnych grach.
}
\description{
Funkcja wywolujaca funkcje \code{gra()} wielokrotnie przy zadanej strategii i zwracajaca wektor wynikow
}
\examples{
wyniki <- badaj_gre(strategia_rf, 10)

}
\author{
Agnieszka Ciepielewska, Marek Wawreniuk, Pawel Gorniak
}

