% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_get_every_animal.R
\name{strategia_get_every_animal}
\alias{strategia_get_every_animal}
\title{Strategia Get Every Animal}
\usage{
strategia_get_every_animal(stock_status, max_stock = c(krolik = 60, owca = 24,
  swinia = 20, krowa = 12, kon = 6, maly_pies = 4, duzy_pies = 2))
}
\arguments{
\item{stock_status}{Stan stada gracza przed ruchem}
}
\value{
Stan stada gracza po ruchu
}
\description{
Strategia polega na zakupywaniu najwyzszego zwierzecia na jakie nas stac, ktorego nie mamy. 
Robione jest to w sposob optymalny, poniewaz korzystamy z \code{expected_value()}. Podczas zakupu brane jest
pod uwage to aby nie zerowac jednego zwierzecia kosztem innego.
}
\author{
Marek Wawreniuk
}

