% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowka_KukurydzaNaPizzy.R
\name{wizytowka_KukurydzaNaPizzy}
\alias{wizytowka_KukurydzaNaPizzy}
\title{Wizytowka}
\usage{
wizytowka_KukurydzaNaPizzy(strategia, powtorz = 10000, sciezka = NULL)
}
\arguments{
\item{strategia}{Strategia, dla ktorej na byc wygenerowana wizytowka}

\item{powtorz}{Liczba powtorzen gier}

\item{sciezka}{Sciezka do pliku, gdzie ma sie zapisac wizytowka}
}
\value{
KukurydzaNaPizzy_Wizytowka_strategia.pdf Wizytowka, pokazujaca statystyki na
temat strategii oraz jej dzialanie. Zostaje ona zapisana w folderze \code{inst} wewnatrz
pakietu
}
\description{
Funkcja \code{wizytowka_KukurydzaNaPizzy} generuje wizytowke wybranej
strategii w formacie pdf
}
\author{
Hanna Kranas, Alicja Gosiewska, Agnieszka Ciepielewska
}

